--- app/Livewire/OrderConsumerShow.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace App\Livewire;
-
-use Livewire\Component;
-
-class OrderConsumerShow extends Component
-{
-    public function render()
-    {
-        return view('livewire.order-consumer-show');
-    }
-}
