+++ resources/views/livewire/blocks/total-orders.blade.php
