+++ resources/views/livewire/consumer/product-show.blade.php
