+++ resources/views/livewire/consumer/user-favorites.blade.php
