--- resources/views/livewire/navbar-cart-counter.blade.php
@@ -1,7 +1,0 @@
-<div>
-    @if($cartCount > 0)
-    <span class="badge bg-danger position-absolute top-0 start-100 translate-middle">
-        {{ $cartCount }}
-    </span>
-    @endif
-</div>
