--- resources/views/livewire/navbar-favorite-counter.blade.php
@@ -1,7 +1,0 @@
-<div>
-    @if($favoritesCount > 0)
-    <span class="badge bg-danger position-absolute top-0 start-100 translate-middle">
-        {{ $favoritesCount }}
-    </span>
-    @endif
-</div>
