--- resources/views/livewire/order-consumer-show.blade.php
@@ -1,3 +1,0 @@
-<div>
-    {{-- Knowing others is intelligence; knowing yourself is true wisdom. --}}
-</div>
