+++ resources/views/user/consumer/product/view.blade.php
@@ -80,7 +80,7 @@
     @include('user.includes.navbar-consumer')
     <div class="main-content-wrapper">
         <!-- All your main page content goes here -->
-        @livewire('product-view', ['productId' => $product->id])
+        @livewire('consumer.product-view', ['productId' => $product->id])
     </div>
 @endsection
